+++ frontend/src/metabase/visualizations/lib/settings/graph.js
@@ -24,7 +24,10 @@ import {
 } from "metabase/visualizations";
 
 import { ChartSettingOrderedSimple } from "metabase/visualizations/components/settings/ChartSettingOrderedSimple";
-import { getDefaultStackingValue } from "metabase/visualizations/shared/settings/cartesian-chart";
+import {
+  getDefaultStackingValue,
+  getSeriesOrderVisibilitySettings,
+} from "metabase/visualizations/shared/settings/cartesian-chart";
 import {
   isDimension,
   isMetric,
@@ -179,54 +182,7 @@ export const GRAPH_DATA_SETTINGS = {
 
     getValue: (series, settings) => {
       const seriesKeys = series.map(s => keyForSingleSeries(s));
-      const seriesSettings = settings["series_settings"];
-      const seriesColors = settings["series_settings.colors"] || {};
-      const seriesOrder = settings["graph.series_order"];
-      // Because this setting is a read dependency of graph.series_order_dimension, this should
-      // Always be the stored setting, not calculated.
-      const seriesOrderDimension = settings["graph.series_order_dimension"];
-      const currentDimension = settings["graph.dimensions"][1];
-
-      if (currentDimension === undefined) {
-        return [];
-      }
-
-      const generateDefault = keys => {
-        return keys.map(key => ({
-          key,
-          color: seriesColors[key],
-          enabled: true,
-          name: seriesSettings[key]?.title || key,
-        }));
-      };
-
-      const removeMissingOrder = (keys, order) =>
-        order.filter(o => keys.includes(o.key));
-      const newKeys = (keys, order) =>
-        keys.filter(key => !order.find(o => o.key === key));
-
-      if (
-        !seriesOrder ||
-        !_.isArray(seriesOrder) ||
-        !seriesOrder.every(
-          order =>
-            order.key !== undefined &&
-            order.name !== undefined &&
-            order.color !== undefined,
-        ) ||
-        seriesOrderDimension !== currentDimension
-      ) {
-        return generateDefault(seriesKeys);
-      }
-
-      return [
-        ...removeMissingOrder(seriesKeys, seriesOrder),
-        ...generateDefault(newKeys(seriesKeys, seriesOrder)),
-      ].map(item => ({
-        ...item,
-        name: seriesSettings[item.key]?.title || item.key,
-        color: seriesColors[item.key],
-      }));
+      return getSeriesOrderVisibilitySettings(settings, seriesKeys);
     },
     getHidden: (series, settings) => {
       return (
